Utilities Folder Rearranged and CSV Utility has been added (#6)
* Uninstalled csv, csv-generate and installed export-to-csv package

* .

* "Update scanner.js to import writeFileSync from fs and update createRectangles.js and generateCSV.js"

This commit updates the file `scanner.js` to import the `writeFileSync` function from the `fs` module directly. Additionally, it makes changes to `createRectangles.js` and `generateCSV.js` files, but the changes are not complete and the functionality of these files is not fully working. The commit message describes the changes as they are and does not make any interpretations.

* "Generate CSV from data array and return CSV buffer"

- Change variable names from `csvYo` to `csvStream` and `csvBuffer`
- Use `csvStream` to generate CSV and convert it to a buffer using `asString()` function.

* Delete utility files and generate CSV with mock data

Deletes `createRectangles.js`, `croppedImage.png`, `eng.traineddata`, `preProcessImage.js`, and `processedImage.png`.
Adds a new line to `generateCSV.js` that generates and saves a CSV file with mock data.
Modifies `out.csv` to include additional columns and data.